% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distri_Sim_functions.R
\name{baysnreg}
\alias{baysnreg}
\title{Bayesian Normal Regression}
\usage{
baysnreg(dat, useW = F, xvec, ndrs = NULL, seed = NULL)
}
\arguments{
\item{dat}{data frame with the first column being the dependent variable}

\item{useW}{a logical indicating if weights are to be used; default \code{useW=F}}

\item{xvec}{predictor vector with respect to which weights are taken}

\item{ndrs}{number of draws of coefficient and variance terms, default: \code{ndrs}=10000}

\item{seed}{the seed set for all random operations in the function, default \code{seed}=1}
}
\value{
beta an ndrs x p matrix of coefficient draws, p is the number of OLS coefficients

sig a vector of variance (sigma squared) draws
}
\description{
This function computes the bayesian normal regression with the option of using weights
}
