% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distri_Sim_functions.R
\name{BB_CBOM}
\alias{BB_CBOM}
\title{\code{BB_CBOM} computes Montiel Olea & Plagborg-Moller 2018 Bayesian simultaneous bands.}
\usage{
BB_CBOM(
  DQmat,
  alpha = 0.1,
  method = "ETI",
  qrangeID = NULL,
  tol = .Machine$double.eps^0.25
)
}
\arguments{
\item{DQmat}{G by L matrix of random draws from the joint distribution}

\item{alpha}{nominal level for \eqn{100(1-\alpha)\%} simultaneous coverage}

\item{method}{type of credible interval to calibrate for simultaneous coverage. Defaults to
the equal-tailed interval (ETI); other options include  'HDI', 'BCI' and 'SI'. 
See \link[bayestestR]{ci} for details.}

\item{qrangeID}{an index of the rows of \code{DQmat} to obtain simultaneous coverage for.}

\item{tol}{a tolerance level as stopping rule for the algorithm.}
}
\value{
BB: Lower and upper Bayesian bands

zeta: the calibrated level at which simultaneous coverage is achieved.
}
\description{
\code{BB_CBOM} computes Montiel Olea & Plagborg-Moller 2018 Bayesian simultaneous bands.
}
\examples{
set.seed(1); BB_CBOM(matrix(rnorm(1000),nrow=5))
}
