% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Opt_and_dist_fns.R
\name{ceHARA}
\alias{ceHARA}
\title{Certainty equivalent from HARA class}
\usage{
ceHARA(parm, rho, W0, R1, pi1)
}
\arguments{
\item{parm}{a parameter vector of allocation to risky asset and trading cost as a fraction of 
wealth}

\item{rho}{the coefficient of relative risk aversion}

\item{W0}{current wealth}

\item{R1}{gross return on the risk-less asset}

\item{pi1}{a vector of random draws from the distribution of excess return on the risky asset}
}
\value{
expected utility
}
\description{
This function computes the certainty equivalent using \code{expHARA()}
}
\examples{
set.seed(30); pi1 = rnorm(15000,0.03954741,0.1868175);R1=1.0046; 
ceHARA(parm = c(0.5,0),rho = 1.5,W0=500,R1=R1,pi1=pi1)

}
