% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distri_Sim_functions.R
\name{compute.Ubands}
\alias{compute.Ubands}
\title{\code{compute.Ubands} is a high-level function for computing uniform confidence bands 
according to Chernozhukov et al. 2013}
\usage{
compute.Ubands(Fn, Boot.Fn, n, alpha = 0.05)
}
\arguments{
\item{Fn}{G-length vector of the function of interest}

\item{Boot.Fn}{a \eqn{G \times L} matrix of samples of the function Fn}

\item{n}{sample size of the data used to compute Fn}

\item{alpha}{significance level}
}
\value{
LB.Fn: G-length lower Uniform Confidence band

UB.Fn: G-length upper Uniform Confidence band
}
\description{
\code{compute.Ubands} is a high-level function for computing uniform confidence bands 
according to Chernozhukov et al. 2013
}
\examples{
set.seed(1); 
Ubands=compute.Ubands(Fn=rep(0,5),Boot.Fn=matrix(rnorm(1000),nrow=5),n=100)

}
