% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Opt_and_dist_fns.R
\name{Costfn}
\alias{Costfn}
\title{Trading cost function}
\usage{
Costfn(parm, W0, R1, A, cpars)
}
\arguments{
\item{parm}{a parameter vector of allocation to risky asset and trading cost as a fraction of 
wealth}

\item{W0}{current wealth}

\item{R1}{gross return on the risk-less asset}

\item{A}{current dollar holding in the risky asset}

\item{cpars}{a two by three matrix whose columns are the quadratic cost parameters,
with the first column corresponding to the fixed cost. For proportional costs, set
the first and third columns to zero.}
}
\value{
cost the dollar value of trading cost
}
\description{
Computes borrowing plus transaction costs using a quadratic cost function
}
\examples{
W0 = 500;A=250;R1=1.0046; cpars<- rbind(c(0.005, 0.0025, 0.00025), c(0.001, 0.0005, 0.00005))
Costfn(parm = c(0.5,0),W0=W0,R1=R1,A=A,cpars = cpars)

}
