% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distri_Sim_functions.R
\name{getprobs}
\alias{getprobs}
\title{Get density values}
\usage{
getprobs(vec, val = NULL, type = "probability")
}
\arguments{
\item{vec}{continuous distribution whose random draws are elements of vec}

\item{val}{a vector of values in the support of vec for which to get pdfs}

\item{type}{the "probability" or "density" that is preferred as output}
}
\value{
mval a vector of probability weights of each element of vec
}
\description{
Get density values for a continuous distribution whose random draws are the vector 
\code{vec}
}
\examples{
set.seed(40); v = rnorm(1000)**2; plot(density(v)); # chi-square distributed
# obtain pdfs at a vector of values x: 
x=c(0,2,4); dn<- getprobs(v,val=x); points(dn$x, dn$y, col = "red");dn

}
