% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCovReg.R
\name{imlmreg.fit}
\alias{imlmreg.fit}
\title{Generic Linear Integrated Moment Regression}
\usage{
imlmreg.fit(Y, D, X = NULL, Z, Kern = "Euclid")
}
\arguments{
\item{Y}{outcome variable}

\item{D}{matrix of endogenous covariates}

\item{X}{matrix of exogenous covariates.}

\item{Z}{matrix of instruments}

\item{Kern}{type of kernel. See \code{\link{Kern.fun}} for available kernels}
}
\value{
an IV regression object which also contains coefficients, standard errors, etc.
}
\description{
\code{imlmreg.fit} runs a generic linear integrated moment regression allowing for different
kernels.
}
\examples{
## Generate data and run MMD regression
n=200; set.seed(12); D = rnorm(n); er = rchisq(n,df=1)-1; Z=D
D=scale(abs(D))+er/sqrt(2); Y=D+er
summary(imlmreg.fit(Y=Y,D=D,Z=Z))
summary(AER::ivreg(formula = Y ~ D | Z)) #compare to conventional IV regression
}
