% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCovReg.R
\name{imlmreg}
\alias{imlmreg}
\title{MMD Regression}
\usage{
imlmreg(Y, X, Z = X, Kern = "Euclid")
}
\arguments{
\item{Y}{outcome variable}

\item{X}{matrix of covariates.}

\item{Z}{matrix of instruments. Defaults to \code{X}.}

\item{Kern}{type of kernel. See \code{\link{Kern.fun}} for available kernels}
}
\value{
an IV regression object which also contains coefficients, standard errors, etc.
}
\description{
\code{imreg} runs a linear integrated moment regression.
}
\examples{
## Generate data and run MMD regression
n=200; set.seed(12); X = rnorm(n); er = rchisq(n,df=1)-1; Z=X; X=scale(abs(X))+er/sqrt(2)
Y=X+er
reg = imlmreg(Y,X,Z) #run regression
AER::ivreg(formula = Y ~ X | Z) #compare to conventional IV regression
}
