% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCovReg.R
\name{imlmreg2.fit}
\alias{imlmreg2.fit}
\title{Generic Linear Integrated Moment Regression}
\usage{
imlmreg2.fit(Y, X, Z, Kern = "Euclid", vctype = "HC3")
}
\arguments{
\item{Y}{outcome variable}

\item{X}{matrix of covariates.}

\item{Z}{matrix of instruments}

\item{Kern}{type of kernel. See Details for available kernels}

\item{vctype}{type of sandwich covariance matrix (see \link[sandwich]{vcovHC})}
}
\value{
an IV regression object which also contains coefficients, standard errors, etc.
}
\description{
\code{imlmreg2.fit} runs a generic linear integrated moment regression allowing for different
kernels. This variant uses centred instruments in the meat of the sandwich matrix
}
\details{
The \eqn{(i,j)}'th elements of available kernel methods are
\describe{
\item{"Euclid"}{Euclidean distance between two vectors: ||Z_i-Z_j||}
\item{"Gauss.W"}{The weighted Gaussian kernel: exp(-0.5(Z_i-Z_j)'V^{-1}(Z_i-Z_j)) where V is the variance of V}
\item{"Gauss"}{The unweighted Gaussian kernel: exp(-||Z_i-Z_j||^2)}
\item{"DL"}{The kernel of Dominguez & Lobato 2004: \eqn{1/n\sum{l=1}^n I(Z_i\le Z_l)I(Z_j\le Z_l)}}
\item{"Esc6"}{The projected version of the DL in Escanciano 2006.}
\item{"WMD"}{The kernel used in Antoine & Lavergne 2014. See page 60 of paper.}
\item{"WMDF"}{The Fuller (1977)-like modification of the kernel in Antoine & Lavergne 2014. See page 64 of paper.}
}
}
\examples{
## Generate data and run MMD regression
n=200; set.seed(12); X = rnorm(n); er = (rchisq(n,df=1)-1)/sqrt(2); Z=X
X=scale(abs(X))+er/sqrt(2); Y=X+er
summary(imlmreg2.fit(Y=Y,X=X,Z=Z))
summary(AER::ivreg(formula = Y ~ X | Z)) #compare to conventional IV regression
}
