% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCovReg.R
\name{imnlmreg.fit}
\alias{imnlmreg.fit}
\title{Nonlinear ICM-IV Regression}
\usage{
imnlmreg.fit(
  strtpar,
  u.fun,
  Xg.fun,
  Y,
  X,
  Z,
  Intercept = F,
  weights = NULL,
  Kern = "Euclid",
  cluster = NULL,
  clus.est.type = "A"
)
}
\arguments{
\item{strtpar}{vector of starting values; use list() to specify lower and upper bounds for line search
if the parameter is a scalar}

\item{u.fun}{function generating the parameterised error}

\item{Xg.fun}{function generating the derivative of u.fun()}

\item{Y}{outcome variable}

\item{X}{matrix of covariates.}

\item{Z}{matrix of instruments}

\item{Intercept}{logical for include intercept u.fun() is non-linear in the intercept}

\item{weights}{a vector of length \eqn{n} of weights for observations}

\item{Kern}{type of kernel. See Details for available kernels}

\item{cluster}{vector of length \eqn{n} with cluster assignments of observations.}

\item{clus.est.type}{options are "A" and "B". "A" sets \eqn{K(Z_i,Z_j)=0} for \eqn{i,j} in the
same cluster while option "B" only does so for \eqn{i=j}.}
}
\value{
an IV regression object which also contains coefficients, standard errors, etc.
}
\description{
\code{imnlmreg.fit} runs a generic possibly non-linear integrated moment regression allowing 
for different kernels. This variant uses centred instruments in the meat of the sandwich matrix
}
\details{
The \eqn{(i,j)}'th elements of available kernel methods are
\describe{
\item{"Euclid"}{Euclidean distance between two vectors: ||Z_i-Z_j||}
\item{"Gauss.W"}{The weighted Gaussian kernel: exp(-0.5(Z_i-Z_j)'V^{-1}(Z_i-Z_j)) where V is the variance of V}
\item{"Gauss"}{The unweighted Gaussian kernel: exp(-||Z_i-Z_j||^2)}
\item{"DL"}{The kernel of Dominguez & Lobato 2004: \eqn{1/n\sum{l=1}^n I(Z_i\le Z_l)I(Z_j\le Z_l)}}
\item{"Esc6"}{The projected version of the DL in Escanciano 2006.}
\item{"WMD"}{The kernel used in Antoine & Lavergne 2014. See page 60 of paper.}
\item{"WMDF"}{The Fuller (1977)-like modification of the kernel in Antoine & Lavergne 2014. See page 64 of paper.}
}
}
\examples{
## Generate data and run regression
n=200; set.seed(12); X = rnorm(n,1,1); er = (rchisq(n,df=1)-1)/sqrt(2); Z=X
Y = X*(5/4)^2 + (5/4)*(X^2)
u.fun = function(theta,Y,X) Y - X*(theta^2) - (X^2)*theta
Xg.fun = function(theta,X) -2*X*theta - X^2
(imnlmreg.fit(list(a=-3.0,b=3.0),u.fun,Xg.fun,Y,X,Z))
}
