% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distri_Sim_functions.R
\name{IndepMH}
\alias{IndepMH}
\title{A Generic Independence Metropolis-Hastings Algorithm}
\usage{
IndepMH(start = NULL, posterior = NULL, ..., propob = NULL, scale = 1.5,
  iter = 15000, burn = 1000)
}
\arguments{
\item{start}{starting values of parameters for the MH algorithm.
It is automatically generated from the normal proposal distribution but the user can also specify.}

\item{posterior}{the posterior distribution function.}

\item{...}{additional arguments to the posterior function
Should take parameter input of the same length as \code{start} or \code{propob$mode}}

\item{propob}{a list of mode and variance-covariance matrix of the normal proposal distribution. 
Save list as propob=list(mode=mode,var=variance-covariance)}

\item{scale}{a value multiplied by \code{propob$var} in order to adjust the proposal distribution.
The default is \code{1.5} but the user can adjust it until a satisfactory acceptance rate is obtained.}

\item{iter}{number of random draws desired (default: 15000)}

\item{burn}{burn-in period for the Random Walk MH algorithm (default: 1000)}
}
\value{
val a list of matrix of draws pardraws and the acceptance rate
}
\description{
\code{IndepMH} computes random draws of parameters using a normal proposal distribution.
This function implements a generic form of \code{IndepMH} from the package \code{bayesdistreg}
}
