% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCovReg.R
\name{Kern.fun}
\alias{Kern.fun}
\title{Construction of n x n Kernel Matrices}
\usage{
Kern.fun(Z, Kern = "Euclid")
}
\arguments{
\item{Z}{n x p matrix of instrumental variables}

\item{Kern}{type of kernel desired.}
}
\value{
the MDD coefficient
}
\description{
\code{Kern.fun} computes the n x n matrix of kernels used in the integrated moment class
of estimators.
}
\details{
The \eqn{(i,j)}'th elements of available kernel methods are
\describe{
\item{"Euclid"}{Euclidean distance between two vectors: ||Z_i-Z_j||} 
\item{"Gauss.W"}{The weighted Gaussian kernel: exp(-0.5(Z_i-Z_j)'V^{-1}(Z_i-Z_j)) where V is the variance of V} 
\item{"Gauss"}{The unweighted Gaussian kernel: exp(-||Z_i-Z_j||^2)}
\item{"DL"}{The kernel of Dominguez & Lobato 2004: \eqn{1/n\sum{l=1}^n I(Z_i\le Z_l)I(Z_j\le Z_l)}}
}
}
\examples{
set.seed(12); X = rnorm(5); Z=cbind(X,abs(X)); Kern.fun(Z,"DL")
}
