% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCovReg.R
\name{Kern.fun_Esc6}
\alias{Kern.fun_Esc6}
\title{Kernel matrix Esc6}
\usage{
Kern.fun_Esc6(Z)
}
\arguments{
\item{Z}{n by pz matrix of instruments Z}
}
\value{
n by n kernel matrix
}
\description{
This a wrapper for a C function that computes the Kernel matrix of Escanciano 2006.
}
\examples{
set.seed(12); Z = rnorm(5); Z=cbind(Z,abs(Z))
Kern.fun_Esc6(Z)
}
