% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCovReg.R
\name{MDD}
\alias{MDD}
\title{Martingale Difference Divergence}
\usage{
MDD(U, Z)
}
\arguments{
\item{U}{the univariate variable}

\item{Z}{the possibly conditioning multivariate variable}
}
\value{
the MDD coefficient
}
\description{
\code{MDD} computes the Martingale Difference Divergence between univariate U and 
possibly multivariate Z. The MDD measures the mean dependence of U on Z.
}
\examples{
set.seed(12); X = rnorm(200); MDD(X,X^2/sqrt(2))
set.seed(12); Y = rchisq(200,1)/sqrt(2); MDD(Y,X)
}
