% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCovReg.R
\name{mdd.t_test}
\alias{mdd.t_test}
\title{An MDD-based t-Test of Mean Independence}
\usage{
mdd.t_test(Y, Z)
}
\arguments{
\item{Y}{univariate variable}

\item{Z}{possibly multivariate variable}
}
\value{
t-statistic, standard error, and p-value of test
}
\description{
\code{mdd.t_test} tests the mean independence of Y conditional on Z using a two-sided
t-test.
}
\examples{
set.seed(12); X = rnorm(200)
mdd.t_test(X,abs(X)); mdd.t_test(abs(X),X)
}
