% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCovReg.R
\name{MDDM}
\alias{MDDM}
\title{Martingale Difference Divergence Matrix}
\usage{
MDDM(U, Z)
}
\arguments{
\item{U}{the multivariate variable}

\item{Z}{the possibly multivariate conditioning variable. One can also directly supply
an n x n distance matrix}
}
\value{
the MDDM matrix
}
\description{
\code{MDDM} computes the Martingale Difference Divergence Matrix between multivariate U and 
possibly multivariate Z of Lee & Shao 2018.
}
\examples{
set.seed(12); U = rnorm(50); U=cbind(U,abs(U)); MDDM(U,2*pnorm(U[,1]))
MDDM(U,as.matrix(dist(2*pnorm(U[,1]))))
}
