% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCovReg.R
\name{mddtest.boot}
\alias{mddtest.boot}
\title{An MDD-based Test of Mean Independence by Wild Bootstrap}
\usage{
mddtest.boot(U, V, B = 199, wmat = NULL, cl = NULL, cluster = NULL)
}
\arguments{
\item{U}{univariate variable}

\item{V}{possibly multivariate variable}

\item{B}{number of wild bootstrap samples. Defaults to 199}

\item{wmat}{in case the n x B matrix of wild bootstrap weights are supplied by the user.}

\item{cl}{an integer to indicate number of child-processes in pbapply::pbsapply().}

\item{cluster}{vector of length n with cluster ids if cluster-robust wild-bootstrap is
used}
}
\value{
test statistic and p-value
}
\description{
\code{mddtest.boot} tests the mean independence of U conditional on V using the wild
bootstrap following Shao & Zhang 2014.
}
\examples{
set.seed(12); X = rnorm(200)
mddtest.boot(X,abs(X)); mddtest.boot(abs(X),X)
}
