% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCovReg.R
\name{mddtest.perm}
\alias{mddtest.perm}
\title{An MDD-based Test of Mean Independence by Permutation Bootstrap}
\usage{
mddtest.perm(U, V, B = 199, PBmat = NULL, cl = NULL)
}
\arguments{
\item{U}{univariate variable}

\item{V}{possibly multivariate variable}

\item{B}{number of wild bootstrap samples. Defaults to 199}

\item{PBmat}{in case the n x B matrix of permuted ids is supplied by the user.}

\item{cl}{an integer to indicate number of child-processes in pbapply::pbsapply().
used}
}
\value{
test statistic and p-value
}
\description{
\code{mddtest.boot} tests the mean independence of U conditional on V using the wild
bootstrap following Shao & Zhang 2014.
}
\examples{
set.seed(12); X = rnorm(200)
mddtest.perm(X,abs(X)); mddtest.perm(abs(X),X)
}
