% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCovReg.R
\name{mdep.nl}
\alias{mdep.nl}
\title{MDep objective function for non-linear models}
\usage{
mdep.nl(theta, U.fun, Z = NULL, Z.m = NULL, sc = 1)
}
\arguments{
\item{theta}{parameter vector}

\item{U.fun}{user-written function of the disturbance function; evaluates to an n x 1 vector}

\item{Z}{n x pz matrix of instruments; should be supplied in case Z.m is null}

\item{Z.m}{Euclidean matrix from instrument matrix Z - optional}

\item{sc}{-  adjusts the objective function - positive for minimisation, 
negative for maximisation}
}
\value{
function value
}
\description{
\code{mdep.nl} computes the objective function of the MDep estimator 
for a non-linear model
}
\examples{
## Generate data and run MMD regression
n=50; set.seed(12); X = rnorm(n); Y = X + (X^2-1)/sqrt(2)
U.fun=function(theta) Y^2-2*Y*(X*theta)-(X*theta)^2
mdep.nl=function(1,U.fun,X)

}
