% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCovReg.R
\name{mindep.boot}
\alias{mindep.boot}
\title{An ICM Test of Mean Independence by Wild Bootstrap}
\usage{
mindep.boot(
  U,
  V,
  B = 199,
  Kern = "Gauss",
  wmat = NULL,
  cl = NULL,
  cluster = NULL
)
}
\arguments{
\item{U}{univariate variable}

\item{V}{possibly multivariate variable}

\item{B}{number of wild bootstrap samples. Defaults to 199}

\item{Kern}{type of kernel. See Details for available kernels}

\item{wmat}{in case the n x B matrix of wild bootstrap weights are supplied by the user.}

\item{cl}{an integer to indicate number of child-processes in pbapply::pbsapply().}

\item{cluster}{vector of length n with cluster ids if cluster-robust wild-bootstrap is
used}
}
\value{
test statistic and p-value
}
\description{
\code{mindep.boot} tests the mean independence of U conditional on V using the wild
bootstrap
}
\examples{
set.seed(12); X = rnorm(200)
mindep.boot(X,abs(X)); mindep.boot(abs(X),X)
}
