% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCovReg.R
\name{mmd.lmspec_test}
\alias{mmd.lmspec_test}
\title{MMD Specification Test}
\usage{
mmd.lmspec_test(mmd.Obj, B = 199, wmat = NULL, cl = NULL, cluster = NULL)
}
\arguments{
\item{mmd.Obj}{is an MMD regression output from mmdreg()}

\item{B}{number of wild bootstrap samples. Defaults to 199}

\item{wmat}{in case the n x B matrix of wild bootstrap weights are supplied by the user.}

\item{cl}{an integer to indicate number of child-processes in pbapply::pbsapply().}

\item{cluster}{vector of length n with cluster ids if cluster-robust wild-bootstrap is
used}
}
\value{
the p-value, test statistic,
}
\description{
\code{mmd.lmspec_test} conducts the Su \& Zheng 2017 specification test a linear model
estimated with the MMD estimator.
}
\examples{
## Generate data and run MMD regression
n=200; set.seed(12); X = rnorm(n); er = rchisq(n,df=1)-1; Z=X; X=scale(abs(X))+er/sqrt(2)
Y=X+er; reg1 = mmdreg(Y,X,Z); reg2 = mmdreg(Y,X,X) #run regression
mmd.lmspec_test(reg1,B=99); mmd.lmspec_test(reg2,B=99) #test under the null and the alternative
## MMD coefficients, standard errors, and t-statistics

}
