% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCovReg.R
\name{mmdlmrelv.b_test}
\alias{mmdlmrelv.b_test}
\title{MMD Relevance Test}
\usage{
mmdlmrelv.b_test(
  X1,
  X2 = NULL,
  Z,
  B = 199,
  wmat = NULL,
  cl = NULL,
  cluster = NULL
)
}
\arguments{
\item{X1}{an n x p1 matrix of endogenous covariates}

\item{X2}{an n x p2 matrix of exogenous covariates}

\item{Z}{an n x pz matrix of instruments}

\item{B}{number of wild bootstrap samples. Defaults to 199}

\item{wmat}{in case the n x B matrix of wild bootstrap weights are supplied by the user.}

\item{cl}{an integer to indicate number of child-processes in pbapply::pbsapply().}

\item{cluster}{vector of length n with cluster ids if cluster-robust wild-bootstrap is
used}
}
\value{
the p-value, test statistic,
}
\description{
\code{mmdlmrelv.b_test} conducts the Su \& Zheng 2017 specification test on a linear model
estimated with the MMD estimator. It is based on the wild bootstrap.
}
\examples{
n=200; set.seed(12); X = rnorm(n); er = (rchisq(n,df=1)-1)/sqrt(2)
X1=(X+er)/sqrt(2);mmdlmrelv.b_test(X1,abs(X),X^3);mmdlmrelv.b_test(X1,X,X^3)

}
