% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCovReg.R
\name{mmdreg.fit}
\alias{mmdreg.fit}
\title{MMD Regression}
\usage{
mmdreg.fit(Y, X, Z = X, cl = NULL)
}
\arguments{
\item{Y}{outcome variable}

\item{X}{matrix of covariates.}

\item{Z}{matrix of instruments. Defaults to \code{X}.}

\item{cl}{number of clusters to pass to \code{pbsapply()}. This is only advised in large samples.}
}
\value{
an IV regression object which also contains coefficients, standard errors, etc. The
standard errors are computed based on a U-Statistics characterisation of the influence function
}
\description{
\code{mmdreg.fit} runs a linear minimum mean dependence (MMD) regression.
}
\examples{
## Generate data and run MMD regression
n=200; set.seed(12); X = rnorm(n); er = rchisq(n,df=1)-1; Z=X; X=scale(abs(X))+er/sqrt(2)
Y=X+er
reg = mmdreg.fit(Y,X,Z) #run regression
## MMD coefficients, standard errors, and t-statistics
reg$MMD_coefficients; reg$MMD_SE; reg$MMD_tstat
}
