% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Functions.R
\name{modslct}
\alias{modslct}
\title{Model Selection}
\usage{
modslct(daten, sbset = NULL)
}
\arguments{
\item{daten}{the data frame of covariates.}

\item{sbset}{if specified as a number from 1 to number of columns of \code{daten}, the matrix
returned corresponds to selections with at least \code{sbset} covariates. If a negative number
from -1 to negative the number of columns, the matrix returned corresponds to selections with
at most \code{sbset} covariates.}
}
\value{
gridm the matrix grid whose rows correspond to inclusion (1) and exclusion (0) of
covariates.
}
\description{
This function generates a best subset selection matrix of zeros and ones corresponding to
inclusion/exclusion of covariates in the data frame \code{daten}.
}
\examples{
gz<- matrix(rnorm(20),ncol = 4); grd<- modslct(gz)
gz[,which(grd[8,]==1)] ## Use 8th subset
(grd1<- modslct(gz,3));(grd1<- modslct(gz,-2))

}
