% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Opt_and_dist_fns.R
\name{optu}
\alias{optu}
\title{Generate distribution of decision parameters}
\usage{
optu(start, rho, W0, R1, pi1, A, cpars, ub, uth, scale = 1.1,
  iter = 15000, burn = 1000)
}
\arguments{
\item{start}{suggested starting values}

\item{rho}{coefficient of relative risk aversion}

\item{W0}{current wealth}

\item{R1}{gross return on the risk-less asset}

\item{pi1}{a vector of random draws from the distribution of excess return on the risky asset}

\item{A}{current dollar holding in the risky asset}

\item{cpars}{a two by three matrix whose columns are the quadratic cost parameters,
with the first column corresponding to the fixed cost. For proportional costs, set
the first and third columns to zero.}

\item{ub}{upper bound on borrowing ; a fraction of wealth}

\item{uth}{upper bound on cost fraction parm[2]}

\item{scale}{a value multiplied by \code{propob$var} in order to adjust the proposal distribution.
The default is \code{1.1} but the user can adjust it until a satisfactory acceptance rate is obtained.}

\item{iter}{number of random draws desired (default: 15000)}

\item{burn}{burn-in period for the Random Walk MH algorithm (default: 1000)}
}
\value{
par parameter values from constrained optimisation

var variance-covariance matrix of parameters

pardraws matrix of parameter draws

postvals distribution of utility obtained from \code{pardraws}

AcceptRatio the acceptance ratio
}
\description{
This function does constrained HARA utility maximisation and outputs optimal parameter values and
the joint distribution of parameters.
}
\examples{
## dj<- optu(start = c(0.5,0),rho = rho,W0=W0,R1=R1,pi1 = pi1,A=A,cpars = cpars,ub=ub,uth = uth)

}
