% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Functions.R
\name{parlply}
\alias{parlply}
\title{Parallel compute}
\usage{
parlply(X, fn, type = "PSOCK", nc = 1, ...)
}
\arguments{
\item{X}{a vector as used in lapply() family of functions}

\item{fn}{the function to be run in parallel}

\item{type}{the type of parallel computation; \code{"PSOCK"} which is compatible with all operating
systems or \code{"FORK"} which is only compatible on Mac/Linux platforms}

\item{nc}{number of clusters to use}

\item{...}{additional inputs for the function \code{fn}.}
}
\value{
val output in a simplified form: vector or matrix
}
\description{
This function takes a function and margin and parallel computes, returning output in a list.
}
