% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distri_Sim_functions.R
\name{predbreg}
\alias{predbreg}
\title{Bayesian Normal Regression}
\usage{
predbreg(dat, xvec, useW = F, ndrs = NULL, seed = NULL)
}
\arguments{
\item{dat}{data frame with the first column being the dependent variable}

\item{xvec}{predictor vector for which a predictive distribution is sought}

\item{useW}{a logical indicating if weights are to be used; default \code{useW=F}}

\item{ndrs}{number of draws of coefficient and variance terms, default: \code{ndrs}=10000}

\item{seed}{the seed set for all random operations in the function, default \code{seed}=1}
}
\value{
beta an ndrs x p matrix of coefficient draws, p is the number of coefficients

ydist draws from the predictive distribution
}
\description{
This function computes the bayesian normal regression with the option of using weights
}
\examples{
Ddat<- bayesprdopt::DAT_ALL[1:75,]; dat<- Ddat[-20,]; xvec=Ddat[20,-1];
dd<- predbreg(dat = dat,xvec = xvec); plot(density(dd$ydist)); plot(density(dd$beta[,3]))
## A weighted version with non-default number of draws
dw<- predbreg(dat = dat,xvec = xvec,useW = TRUE,ndrs = 9000); plot(density(dw$ydist))

}
