% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distri_Sim_functions.R
\name{predbregspc}
\alias{predbregspc}
\title{Bayesian predictive distribution with supervised principal components}
\usage{
predbregspc(dat, xvec, useW = F, ndrs = NULL, seed = NULL, lev1 = 0.1,
  lev2 = 0.95, maxc = 3, minft = 6)
}
\arguments{
\item{dat}{data frame with the first column being the dependent variable}

\item{xvec}{predictor vector for which a predictive distribution is sought}

\item{useW}{a logical indicating if weights are to be used; default \code{useW=F}}

\item{ndrs}{number of draws of coefficient and variance terms, default: \code{ndrs}=10000}

\item{seed}{the seed set for all random operations in the function, default \code{seed}=1}

\item{mods}{a vector of indexes of models on the \code{grid} from \code{modslct} to use. The
default computes predictive distributions for all models.}

\item{sbset}{minimum number of covariates in each model.
Use negative to have maximum number of covariates in each model.}
}
\value{
mat a matrix, with columns corresponding to predictive distributions
}
\description{
Bayesian predictive distribution with supervised principal components
}
\examples{
Ddat<- bayesprdopt::DAT_ALL[1:75,]; dat<- Ddat[-20,]; xvec=Ddat[20,-1];
yspc<- predbregspc(dat,xvec); plot(density(yspc))

}
