% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distri_Sim_functions.R
\name{predist}
\alias{predist}
\title{Get predictive distribution}
\usage{
predist(baysobj, xvec)
}
\arguments{
\item{baysobj}{object from \code{baysnreg} which has matrix of coefficient and \code{sigma} draws.}

\item{xvec}{the predictor vector for which the predictive distribution is sought}
}
\value{
ydist a vector representing the predictive distribution.
}
\description{
This function obtains random draws from the predictive distribution
}
