% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Supervised_principal_components.R
\name{sfeat}
\alias{sfeat}
\title{Significant feature selection}
\usage{
sfeat(y, x, lev1 = 0.1, minft = 6)
}
\arguments{
\item{y}{predictand -  a vector of length n}

\item{x}{n x p matrix, with possibly p>n}

\item{lev1}{maximum level of significance to consider; (default 0.1)}

\item{minft}{minimum number of features to select. If the number of selected features is
less than \code{minft}, return the entire design matrix \code{x}; (default 6)}
}
\value{
xslct a subset matrix of \code{x} whose columns comprises selected features

slctfeat a vector of indices of selected features
}
\description{
\code{sfeat} produces a subset a matrix x with features significant at at most \code{lev1}
level of significance.
}
