% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCovReg.R
\name{speclmb.test}
\alias{speclmb.test}
\title{ICM Specification Test for linear models}
\usage{
speclmb.test(
  reg.Obj,
  Kern = "Euclid",
  B = 199,
  wmat = NULL,
  cl = NULL,
  cluster = NULL
)
}
\arguments{
\item{reg.Obj}{is a regression output}

\item{Kern}{type of kernel desired.}

\item{B}{number of wild bootstrap samples. Defaults to 199}

\item{wmat}{in case the n x B matrix of wild bootstrap weights are supplied by the user.}

\item{cl}{an integer to indicate number of child-processes in pbapply::pbsapply().}

\item{cluster}{vector of length n with cluster ids if cluster-robust wild-bootstrap is
used}
}
\value{
the p-value, test statistic,
}
\description{
\code{speclmb.test} conducts the Su \& Zheng 2017 specification test on a linear model
estimated with the MMD estimator. It is based on the wild bootstrap.
}
\examples{
## Generate data and run MMD regression
n=100; set.seed(12); X = rnorm(n); er = rnorm(n)
X=scale(abs(X))+er/sqrt(2); Y=X+er
speclmb.test(imlmreg.fit(Y,X,X)) #null
speclmb.test(imlmreg.fit(Y,abs(X),X)) #alternative

}
