% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stochastic_Dominance_Tests.R
\name{stochdom}
\alias{stochdom}
\title{Stochastic Dominance}
\usage{
stochdom(X, Y, order = 1, type = "Bootstrap", B = 2000)
}
\arguments{
\item{X}{vector of first variable}

\item{Y}{vector of second variable}

\item{order}{for order of test "equality", 1, 2, 3, "all"}

\item{type}{"Bootstrap" following Abadie (2002) or "semipar" following Anderson (1996)}
}
\value{
val an ndrs x p matrix of coefficient draws, p is the number of OLS coefficients
}
\description{
This function tests the stochastic dominance of two variables \code{X} and \code{Y}
}
\examples{
# Not run
# set.seed(30); X=rnorm(200,mean = 0.2,sd=2); set.seed(30); Y=rnorm(400,0.24,sd=1)
# stochdom(Y,X,order = 1); stochdom(X,Y,order = 2)
# set.seed(30); X=rchisq(1000,df=14); set.seed(30); Y=rnorm(278,mean=6,sd=2)
# stochdom(Y,X,order = 1);stochdom(X,Y,order = 1); 
# stochdom(X,Y,order = 2); stochdom(Y,X,order = 2)

}
