% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distri_Sim_functions.R
\name{sup_t_Band.Fn}
\alias{sup_t_Band.Fn}
\title{\code{sup_t_Band.Fn} constructs the sup-t uniform band}
\usage{
sup_t_Band.Fn(Fn, DQmat, Sigv, n, alpha = 0.05)
}
\arguments{
\item{Fn}{G-length vector of the function of interest}

\item{DQmat}{\eqn{G\times L} matrix of mean-zero random draws from the (joint) distribution}

\item{Sigv}{the \eqn{G\times 1} vector of standard deviations corresponding to estimates Fn}

\item{n}{sample size of the data used to compute Fn}

\item{alpha}{significance level}
}
\value{
LB.Fn: G-length lower Uniform Confidence band

UB.Fn: G-length upper Uniform Confidence band
}
\description{
\code{sup_t_Band.Fn} constructs the sup-t uniform band
}
\examples{
set.seed(1); DQmat=matrix(rnorm(1000),nrow=5)
sup_t_Band.Fn(Fn=(1:5)/5,DQmat=DQmat,Sigv=rep(1,5),n=100)

}
