% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distri_Sim_functions.R
\name{sup_t_Band.LL}
\alias{sup_t_Band.LL}
\title{\code{sup_t_Band.LL} constructs the sup-t uniform band following Li & Liao (2020)}
\usage{
sup_t_Band.LL(
  reg.objs,
  lambda,
  L = NULL,
  norm.mat = NULL,
  n,
  alpha = 0.05,
  seed = NULL
)
}
\arguments{
\item{reg.objs}{G-length list of regression objects}

\item{lambda}{\eqn{\lambda'\beta} is the functional of interest}

\item{L}{the number of variates to use in simulating the distribution of sup-t. Should
be supplied or norm.mat should be supplied}

\item{norm.mat}{\eqn{k \times L} matrix of standard normal variates used
in computing the bands}

\item{n}{sample size of the data used to compute reg.objs}

\item{alpha}{significance level}

\item{seed}{seed specified if norm.mat=NULL}
}
\value{
Fn: G-length vector of the function estimate

LB.Fn: G-length lower Uniform Confidence band

UB.Fn: G-length upper Uniform Confidence band
}
\description{
\code{sup_t_Band.LL} constructs the sup-t uniform band following Li & Liao (2020)
}
\examples{
set.seed(1); X = rnorm(250,mean=(1:250)/250)
reg.objs = lapply(seq(0.2,0.8,by=0.05), function(t) lm(I(X+t)~1))
sup_t_Band.LL(reg.objs,lambda=1,L=500,n=250,seed=1)

}
