% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Supervised_principal_components.R
\name{supc}
\alias{supc}
\title{Supervised principal component}
\usage{
supc(dat, xvec, lev1 = 0.1, lev2 = 0.95, maxc = 3, minft = 6)
}
\arguments{
\item{dat}{data set with first column being the predictand/dependent variable}

\item{xvec}{predictor vector of interest. Its supervised component form will be the first row of the matrix returned}

\item{lev1}{lev1 maximum level of significance to consider; (default 0.1)}

\item{lev2}{highest fraction of cumulative variation of supervised principal components: default 0.95}

\item{maxc}{maximum number of supervised principal components to return. Note, this option overrides
the \code{lev} option if \code{maxc} is set to a number smaller than the one dictated by \code{lev2}}

\item{minft}{minimum number of features to select. If the number of selected features is
less than \code{minft}, return the entire design matrix \code{x}; (default 6)}
}
\value{
slctcomp  a matrix of maxc supervised principal components.
}
\description{
This function conducts principal components on a subset matrix \code{xx}
}
\examples{
dat<- prdtrade::DAT_ALL[1:75,]; dat<- Ddat[-20,]; xvec=Ddat[20,-1];
spc<- supc(dat = dat,xvec=xvec)

}
