% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Functions.R
\name{testSR}
\alias{testSR}
\title{Probability value of Sharpe Ratio}
\usage{
testSR(ydist = NULL, yt = 0, yhat = NULL, ysd = NULL)
}
\arguments{
\item{ydist}{a vector of draws from the distribution of excess returns}

\item{yt}{the value of comparison, i.e. we test if expected excess returns is equal to yt; (default; yt=0)}

\item{yhat}{expected excess returns; provide this option if you do not have the distribution \code{ydist}}

\item{ysd}{the volatility of excess returns; provide this input if you do not have the distribution \code{ydist}}
}
\value{
pval the probability value
}
\description{
This function computes the probaility value of the Sharpe ratio
}
\examples{
d<- rnorm(5000,-0.023,1.3); testSR(d); testSR(yt=0.5,yhat=0.11,ysd=1.01)

}
