% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distri_Sim_functions.R
\name{UBand.Fn}
\alias{UBand.Fn}
\title{\code{UBand.Fn} constructs Chernozhukov et al. 2013 Uniform Confidence Bands
using output from \link{Band.Fn.objs}}
\usage{
UBand.Fn(Fn, BDobjs, n, alpha = 0.05)
}
\arguments{
\item{Fn}{G-length vector of the function of interest}

\item{BDobjs}{output object from \link{Band.Fn.objs}}

\item{n}{sample size of the data used to compute Fn}

\item{alpha}{significance level}
}
\value{
LB.Fn: G-length lower Uniform Confidence band

UB.Fn: G-length upper Uniform Confidence band
}
\description{
\code{UBand.Fn} constructs Chernozhukov et al. 2013 Uniform Confidence Bands
using output from \link{Band.Fn.objs}
}
\examples{
set.seed(1); 
BDobjs=Band.Fn.objs(Fn=rep(0,5),Boot.Fn=matrix(rnorm(1000),nrow=5),n=100)
UBand.Fn(Fn=rep(0,5),BDobjs=BDobjs,n=100)

}
