% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCovReg.R
\name{wmat.mammen}
\alias{wmat.mammen}
\title{Mammen's Wild Bootstrap Weights}
\usage{
wmat.mammen(n, B = 200, seed = NULL, cluster = NULL)
}
\arguments{
\item{n}{sample size}

\item{B}{number of wild bootstrap samples.}

\item{seed}{seed for reproducibility.}

\item{cluster}{vector of length n with cluster ids if cluster-robust wild-bootstrap is
used}
}
\value{
an n by B matrix of two-point weights
}
\description{
\code{wmat.mammen} generates Mammen's two-point weight for the wild bootstrap.
}
