% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCovReg.R
\name{kClassIVreg.fit}
\alias{kClassIVreg.fit}
\title{Generic k-class estimator}
\usage{
kClassIVreg.fit(Y, X, Z, method = "JIVE")
}
\arguments{
\item{Y}{outcome variable}

\item{X}{matrix of covariates.}

\item{Z}{matrix of instruments. Defaults to \code{X}.}

\item{method}{method of the k-class to implement. Defaults to "JIVE".}
}
\value{
an IV regression object which also contains coefficients, standard errors, etc.
}
\description{
\code{kClassIVreg.fit} runs a generic linear IV model of the k-Class
}
\details{
Available methods in the k-Class include
\describe{
\item{"JIVE"}{The Jackknife IV of Angrist et al. 1999} 
\item{"LIML"}{Limited Maximum Likelihood} 
\item{"HLIM"}{The Jackknife Limited Maximum Likelihood of Hausman et al. 2012}
\item{"HFUL"}{The heteroskedasticity robust version of the Fuller (1977) estimator}
}
}
\examples{
## Generate data and run MMD regression
n=200; set.seed(12); X = rnorm(n); er = rchisq(n,df=1)-1; Z=X 
Z=cbind(Z,Z^2,Z^3,Z^4);X=scale(abs(X))+er/sqrt(2); Y=X+er
summary(kClassIVreg.fit(Y=Y,X=X,Z=Z))
summary(AER::ivreg(formula = Y ~ X | Z)) #compare to conventional IV regression
}
