% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distri_Sim_functions.R
\name{sup_t_Band.Fn}
\alias{sup_t_Band.Fn}
\title{\code{sup_t_Band.Fn} constructs the sup-t uniform band}
\usage{
sup_t_Band.Fn(Fn, Sigv, DQmat, n, alpha = 0.05)
}
\arguments{
\item{Fn}{G-length vector of the function of interest}

\item{Sigv}{the \eqn{G \times 1} vector of standard deviations corresponding to 
estimates Fn}

\item{DQmat}{\eqn{G \times L} matrix of standard normal random draws 
from the (joint) distribution of Fn}

\item{n}{sample size of the data used to compute Fn}

\item{alpha}{significance level}
}
\value{
LB.Fn: G-length lower Uniform Confidence band

UB.Fn: G-length upper Uniform Confidence band
}
\description{
\code{sup_t_Band.Fn} constructs the sup-t uniform band
}
\examples{
set.seed(1); DQmat=matrix(rnorm(1000),nrow=5)
sup_t_Band.Fn(Fn=(1:5)/5,Sigv=rep(1,5),DQmat=DQmat,n=100)

}
